Eliminate need for network access
Checking the remote for references would need a lot of code for
authentication and co since there are so many different git auth and
transport implementations and users use this in many different ways.
To keep git-sweep small and easy we rely on core git functionalities
like pruning which removes the local references for remote branches and
indicates for git-sweep that the remote branch is gone.
Therefore git-sweep itself doesn't need network access anymore.